Changes to #ifdef for ncurses and termcap (not finished). These #ifdef's don't always make sense. They need further investigation.
GCC V8 has some new warnings such as -Wstringop-truncation where a string may not
be copied at all (one of the cases occurring here). The other one is when using
snprintf and the total size of the output may be larger than the string can handle.

Most of the changes were to illiminate the warnings and prevent errors.

When starting the edit of these sources I translated '#ifndef TERMCAP' to `#ifdef curses || ncurses'
(not literally correct). These #ifdefs made no sense at all in the original setting, so when changed
still made no sense.

We may need to decide on whether termcap will be a test, or we should require ncurses to be present
for this software to work.